refactor(proactive): extract capability domains
